Spring cleaning for bringup tooling (#3)
* Ignore raw flash images

* Add building and flashing debug variant to makefile

* Use Python interpreter from env/search path

This allows to use a custom version and virtual environments

* Disable ADC control/voltage controller for calibration

The ADC contron/voltage controller overwrites PWM settings. It got
introduced after the calibration scripts and needs to be disable before
any raw PWM settings can take effect.

* Use Ledger's quicklog package from GitHub

The quicklog package on PyPI is not the one you are looking for.

* Use auto device detection in calibration scripts

A single SiliconToaster which got prepared according to the ftdi/ folder
will be detected automatically by SiliconToaster.

* Clean up USB devices used in firmware/makefile

The device-specific port name from my FTDI device sneaked in
accidentally.